ExplorePage + InstalledSnapsPage: use new banners
