+++ lib/app/common/app_finding.dart
@@ -1,0 +1,16 @@
+import 'package:appstream/appstream.dart';
+import 'package:snapd/snapd.dart';
+
+class AppFinding {
+  final Snap? snap;
+  final AppstreamComponent? appstream;
+  final double? rating;
+  final int? totalRatings;
+
+  AppFinding({
+    this.snap,
+    this.appstream,
+    this.rating,
+    this.totalRatings,
+  });
+}
